server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-pptx==0.6.23
Jinja2==3.1.2
pydantic==2.5.2
bcrypt==4.1.2
anthropic==0.42.0
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
beautifulsoup4==4.12.3
perplexityai==0.30.1
